Refactored migration code
