+++ hack/runtime-migrator/internal/migration/migrator.go
@@ -1,4 +1,4 @@
-package runtime
+package migration
 
 import (
 	"context"
