+++ hack/runtime-migrator/internal/migration/migrator_test.go
@@ -1,4 +1,4 @@
-package runtime
+package migration
 
 import (
 	"github.com/stretchr/testify/require"
