+++ hack/runtime-migrator/internal/migration/verifier.go
@@ -1,4 +1,4 @@
-package runtime
+package migration
 
 import (
 	"slices"
